Added to FileIO and added Main.py
File IO can now read a basic track (assuming it is a 'type 1' track) and save it into simple song, track, and note objects. This currently cannot be reconstructed into a full midi file, and does not store all of the fields needed to function correctly.

Also added 'main', a class we can use to test our code and see if it functions as we expect (This may get removed in the final version of the program)

Added a folder for midi samples. Currently contains one song