Feature/sd 9/add file io (#12)
* Added handle_control test

Added testing for the handle_control method in File IO

* Added song constructor test

Added tests for inputs to song constructor

* Increased testing of FileIO

Added more testing of FileIO, and generated printouts of many of our sample songs. These may be helpful when testing and developing to understand the contents of our midi library better

* Added testing for change key

Added test cases for all methods related to changing the key of a song. I used custom C scale and D scale midi files to conduct the testing.

* Refactor larger methods into smaller chunks and add docstrings to FileIO

* Fixed some bugs in the abstracted code

Fixed some bugs in the abstracted methods that prevented track from being initialized.

Removed "check_control_message" method as it would be more difficult that originally thought to abstract this section of code.

Updated comments for the methods in FileIO

Started adding an "equals" method in song that will compare the contents of two song objects for equality

* Delete workspace.xml

This file should be ignored by the .gitignore file, but it does not seem to be getting ignored.

* Improved testing of FileIO

Added an 'equals' method to song that compares two song objects' contents for equality.  Used this method to test fileIO more strongly

* Reverted accidental changes

Reverted accidental changes to 2021FallTeam17-DeHaan.iml

* reverted accidental changes to misc.xml

Co-authored-by: elinewman12 <elis.newman12@gmail.com>
Co-authored-by: Erin Litty <erlitty@ncsu.edu>